Add gh workflows, add basic testing, refactor examples, improve coverage
