+++ umshini/__init__.py
@@ -1,3 +1,5 @@
+from .envs import ALL_ENVIRONMENTS, CLASSIC_GAMES, LLM_GAMES
+
 from .learner import connect, test
 
 __version__ = "0.0.17"
