+++ umshini/learner.py
@@ -1,9 +1,13 @@
 # pyright: reportGeneralTypeIssues=false
+from __future__ import annotations
+
+from typing import Optional
 
 from colorama import Fore, Style
 
 from umshini.example_client import UmshiniTournamentAgent
 from umshini.tournament_client import TestEnv
+from umshini.examples.example_agent import DummyAgent
 
 
 def create_and_run(botname, user_key):
@@ -43,8 +47,10 @@ def connect(environment, botname, user_key, user_policy, debug=False, testing=Fa
     agent.run()
 
 
-def test(environment, user_policy):
-    test_env = TestEnv(environment)
+def test(env_id: str, user_policy: Optional[callable] = None):
+    if user_policy is None:
+        user_policy = DummyAgent(env_id).pol
+    test_env = TestEnv(env_id)
     term = trunc = False
     obs = rew = info = None
     for _ in range(100):
@@ -55,7 +61,7 @@ def test(environment, user_policy):
             if term or trunc:
                 print(
                     Fore.GREEN
-                    + f"Policy has passed verification testing in {environment}"
+                    + f"Policy has passed verification testing in {env_id}"
                 )
                 print(Style.RESET_ALL)
                 break
